requirements.txt
absl-py==0.1.10
alabaster==0.7.9
altgraph==0.15
anaconda-clean==1.0
anaconda-client==1.5.1
anaconda-navigator==1.3.1
argcomplete==1.0.0
astroid==1.4.7
astropy==1.2.1
async==0.6.2
Babel==2.3.4
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
backports.weakref==1.0.post1
beautifulsoup4==4.5.1
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.2
boto==2.42.0
Bottleneck==1.1.0
cdecimal==2.3
certifi==2018.1.18
cffi==1.7.0
chardet==3.0.4
chest==0.2.3
click==6.6
cloudpickle==0.2.1
clyent==1.2.2
colorama==0.3.7
conda==4.4.9
conda-build==2.0.2
configobj==5.0.6
configparser==3.5.0
contextlib2==0.5.3
cryptography==1.5
cx-Freeze==5.1.1
cycler==0.10.0
Cython==0.24.1
cytoolz==0.8.0
dask==0.11.0
datashape==0.5.2
decorator==4.0.10
dill==0.2.5
dis3==0.1.2
dj-database-url==0.4.2
Django==1.11.10
django-heroku==0.2.0
docutils==0.12
dynd==0.7.3.dev1
easydict==1.4
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.6
Flask==0.11.1
Flask-Cors==2.1.2
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
gevent==1.1.2
google-api-python-client==1.6.5
greenlet==0.4.10
grin==1.2.1
gunicorn==19.7.1
h5py==2.6.0
HeapDict==1.0.0
html5lib==0.9999999
httplib2==0.10.3
idna==2.6
imagesize==0.7.1
ipaddress==1.0.16
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
lazy-object-proxy==1.2.1
llvmlite==0.13.0
locket==0.2.0
lxml==3.6.4
macholib==1.9
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.3
mock==2.0.0
mpmath==0.19
multipledispatch==0.4.8
nb-anacondacloud==1.2.0
nb-conda==2.0.0
nb-conda-kernels==2.0.0
nbconvert==4.2.0
nbformat==4.1.0
nbpresent==3.0.2
networkx==1.11
nltk==3.2.1
nose==1.3.7
notebook==4.2.3
numba==0.28.1+0.gfe99fbc.dirty
numexpr==2.6.1
numpy==1.14.0
oauth2client==4.1.2
odo==0.5.0
openpyxl==2.3.2
pandas==0.18.1
partd==0.3.6
path.py==0.0.0
pathlib==1.0.1
pathlib2==2.1.0
patsy==0.4.1
pbr==3.1.1
pefile==2017.11.5
pep8==1.7.0
pew==1.1.2
pexpect==4.0.1
pickleshare==0.7.4
Pillow==3.3.1
pipenv==10.1.0
pkginfo==1.3.2
ply==3.9
prompt-toolkit==1.0.3
protobuf==3.5.1
psutil==4.3.1
psycopg2==2.7.4
ptyprocess==0.5.1
py==1.4.31
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.10.0
pycosat==0.6.3
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.3.0
Pygments==2.1.3
PyInstaller==3.3.1
pylint==1.5.4
pyOpenSSL==16.2.0
pyparsing==2.1.4
PySide==1.2.4
pytesseract==0.2.0
pytest==2.9.2
python-dateutil==2.5.3
pytz==2016.6.1
PyYAML==3.12
pyzmq==15.4.0
QtAwesome==0.3.3
qtconsole==4.2.1
QtPy==1.1.2
redis==2.10.5
requests==2.18.4
rope==0.9.4
rsa==3.4.2
ruamel-yaml===-VERSION
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.18.1
shutilwhich==1.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.4.6
spyder==3.0.0
SQLAlchemy==1.0.13
statsmodels==0.6.1
sympy==1.0
tables==3.2.3.1
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
terminado==0.6
toolz==0.8.0
tornado==4.4.1
traitlets==4.3.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==3.3.1
widgetsnbextension==1.2.6
wrapt==1.10.6
xlrd==1.0.0
XlsxWriter==0.9.3
xlwt==1.1.2
